% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline-interactive-widget.R
\name{inlineOutput}
\alias{inlineOutput}
\title{Create an inline output span}
\usage{
inlineOutput(outputId)
}
\arguments{
\item{outputId}{The output identifier used to update the value from server logic}
}
\value{
A Shiny UI output element configured for inline display
}
\description{
Creates an output element that can be embedded inline within text.
}
\examples{
library(shiny)

ui <- fluidPage(
  p("The result is", inlineOutput("result"))
)

server <- function(input, output) {
  output$result <- renderInline({
    # Will be formatted to one decimal place
    42.123
  })
}

if (interactive()) {
 shinyApp(ui, server)
}
}
