% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline-interactive-widget.R
\name{renderInline}
\alias{renderInline}
\title{Create an inline render function}
\usage{
renderInline(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{Expression to evaluate}

\item{env}{Environment to evaluate in}

\item{quoted}{Whether the expression is quoted}
}
\value{
A Shiny render function that creates inline output elements
}
\description{
Creates a render function for inline outputs that automatically formats numeric
values appropriately (integers without decimals, other numbers to one decimal place).
}
\examples{
library(shiny)

ui <- fluidPage(
  p("The result is", inlineOutput("result"))
)

server <- function(input, output) {
  output$result <- renderInline({
    # Will be formatted to one decimal place
    42.123
  })
}

if (interactive()) {
 shinyApp(ui, server)
}
}
