% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytangle-package.R
\docType{package}
\name{shinytangle-package}
\alias{shinytangle}
\alias{shinytangle-package}
\title{shinytangle: Interactive Inline TangleJS Style Elements for Shiny Applications}
\description{
Provides a set of functions for creating inline numeric inputs and outputs in Shiny applications with drag-to-change functionality inspired by Tangle.js.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless-rpkg/shinytangle}
  \item \url{https://r-pkg.thecoatlessprofessor.com/shinytangle/}
  \item Report bugs at \url{https://github.com/coatless-rpkg/shinytangle/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{james.balamuta@gmail.com}

}
\keyword{internal}
